Folderリクエストを別ファイルに定義 #70
